requirements.txt
twitter-api-client
instagrapi
mysql-connector-python
instaloader 
Pillow